move user messages into api folder
