+++ nanome/api/user/callbacks.py
@@ -1,0 +1,3 @@
+
+def presenter_change(network, arg, request_id):
+    network._on_presenter_change()
